+++ app/(home)/(tabs)/_layout.tsx
@@ -46,7 +46,7 @@ export default function TabsLayout() {
           }}
         />
         <Tabs.Screen
-          name="home"
+          name="index"
           options={{
             title: "Community",
             tabBarIcon: ({ color, focused }) => (
