+++ app/(home)/room/[id].tsx
