--- app/(pages)/[username].tsx
@@ -1,18 +1,0 @@
-import React from "react";
-import { View, Text, Pressable } from "react-native";
-import { router, useLocalSearchParams } from "expo-router";
-
-const UserScreen = () => {
-  const { username } = useLocalSearchParams<{ username: string }>();
-
-  return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text>{username}</Text>
-      <Pressable onPress={() => router.back()}>
-        <Text>Go back</Text>
-      </Pressable>
-    </View>
-  );
-};
-
-export default UserScreen;
